refactor: activeGeneration activityRecord 패키지 분리
