+++ src/main/java/com/tavemakers/surf/domain/activity/dto/acitveGeneration/request/ActiveGenerationPatchReqDTO.java
@@ -1,4 +1,4 @@
-package com.tavemakers.surf.domain.activity.dto.request;
+package com.tavemakers.surf.domain.activity.dto.acitveGeneration.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
