+++ src/main/java/com/tavemakers/surf/domain/activity/dto/activityRecord/request/ActivityRecordPatchReqDTO.java
@@ -1,0 +1,11 @@
+package com.tavemakers.surf.domain.activity.dto.activityRecord.request;
+
+import com.tavemakers.surf.domain.activity.entity.enums.ActivityType;
+
+import java.time.LocalDate;
+
+public record ActivityRecordPatchReqDTO(
+        ActivityType activityType,
+        LocalDate activityDate
+) {
+}
