+++ src/main/java/com/tavemakers/surf/domain/activity/dto/activityRecord/request/ActivityRewardGroupReqDTO.java
@@ -1,5 +1,7 @@
-package com.tavemakers.surf.domain.activity.dto.response;
+package com.tavemakers.surf.domain.activity.dto.activityRecord.request;
 
+import com.tavemakers.surf.domain.activity.dto.activityRecord.response.ActivityTypeCountResDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.response.ActivityTypeGroupCountResDTO;
 import lombok.Builder;
 
 import java.util.List;
