+++ src/main/java/com/tavemakers/surf/domain/activity/dto/activityRecord/response/ActivityTypeCountResDTO.java
@@ -1,4 +1,4 @@
-package com.tavemakers.surf.domain.activity.dto.response;
+package com.tavemakers.surf.domain.activity.dto.activityRecord.response;
 
 import com.tavemakers.surf.domain.activity.entity.enums.ActivityType;
 import lombok.Builder;
