+++ src/main/java/com/tavemakers/surf/domain/activity/dto/activityRecord/response/AdminActivityRecordSliceResDTO.java
@@ -1,20 +1,22 @@
-package com.tavemakers.surf.domain.activity.dto.response;
+package com.tavemakers.surf.domain.activity.dto.activityRecord.response;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import org.springframework.data.domain.Slice;
 
 import java.util.List;
 
 @Builder
-public record ActivityRecordSliceResDTO(
-        List<ActivityRecordResDTO> content,
+public record AdminActivityRecordSliceResDTO(
+        List<AdminActivityRecordResDTO> content,
         int pageNumber,
         int pageSize,
         int numberOfElements,
-        boolean isLast
+        @JsonProperty("isLast") boolean isLast
 ) {
-    public static ActivityRecordSliceResDTO from(Slice<ActivityRecordResDTO> slice) {
-        return ActivityRecordSliceResDTO.builder()
+    /** Slice 래퍼 생성 */
+    public static AdminActivityRecordSliceResDTO from(Slice<AdminActivityRecordResDTO> slice) {
+        return AdminActivityRecordSliceResDTO.builder()
                 .content(slice.getContent())
                 .pageNumber(slice.getNumber())
                 .pageSize(slice.getSize())
