+++ src/main/java/com/tavemakers/surf/domain/activity/service/activeGeneration/ActiveGenerationGetService.java
@@ -1,4 +1,4 @@
-package com.tavemakers.surf.domain.activity.service;
+package com.tavemakers.surf.domain.activity.service.activeGeneration;
 
 
 import com.tavemakers.surf.domain.activity.entity.ActiveGeneration;
