+++ src/main/java/com/tavemakers/surf/domain/activity/usecase/ActivityRecordUsecase.java
@@ -1,23 +1,23 @@
 package com.tavemakers.surf.domain.activity.usecase;
 
-import com.tavemakers.surf.domain.activity.dto.request.ActivityRecordPatchReqDTO;
-import com.tavemakers.surf.domain.activity.dto.request.ActivityRecordReqDTO;
-import com.tavemakers.surf.domain.activity.dto.request.ActivityRecordReqDTOV2;
-import com.tavemakers.surf.domain.activity.dto.response.ActivityCategoryDetailResDTO;
-import com.tavemakers.surf.domain.activity.dto.response.ActivityCategoryResDTO;
-import com.tavemakers.surf.domain.activity.dto.response.ActivityRecordResDTO;
-import com.tavemakers.surf.domain.activity.dto.response.ActivityRecordSliceResDTO;
-import com.tavemakers.surf.domain.activity.dto.response.AdminActivityRecordResDTO;
-import com.tavemakers.surf.domain.activity.dto.response.AdminActivityRecordSliceResDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.request.ActivityRecordPatchReqDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.request.ActivityRecordReqDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.request.ActivityRecordReqDTOV2;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.response.ActivityCategoryDetailResDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.response.ActivityCategoryResDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.response.ActivityRecordResDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.response.ActivityRecordSliceResDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.response.AdminActivityRecordResDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.response.AdminActivityRecordSliceResDTO;
 import com.tavemakers.surf.domain.activity.entity.ActivityRecord;
 import com.tavemakers.surf.domain.activity.entity.enums.ActivityCategory;
 import com.tavemakers.surf.domain.activity.entity.enums.ActivityType;
 import com.tavemakers.surf.domain.activity.entity.enums.ScoreType;
 import com.tavemakers.surf.domain.activity.exception.ActivityRecordAlreadyDeletedException;
-import com.tavemakers.surf.domain.activity.service.ActivityRecordDeleteService;
-import com.tavemakers.surf.domain.activity.service.ActivityRecordGetService;
-import com.tavemakers.surf.domain.activity.service.ActivityRecordCreateService;
-import com.tavemakers.surf.domain.activity.service.ActivityRecordPatchService;
+import com.tavemakers.surf.domain.activity.service.activityRecord.ActivityRecordDeleteService;
+import com.tavemakers.surf.domain.activity.service.activityRecord.ActivityRecordGetService;
+import com.tavemakers.surf.domain.activity.service.activityRecord.ActivityRecordCreateService;
+import com.tavemakers.surf.domain.activity.service.activityRecord.ActivityRecordPatchService;
 import com.tavemakers.surf.domain.score.entity.PersonalActivityScore;
 import com.tavemakers.surf.domain.score.service.PersonalScoreGetService;
 import com.tavemakers.surf.domain.score.utils.ScoreCalculator;
