+++ src/main/java/com/tavemakers/surf/domain/score/controller/ScoreRankingGetController.java
@@ -1,6 +1,6 @@
 package com.tavemakers.surf.domain.score.controller;
 
-import com.tavemakers.surf.domain.activity.dto.response.TeamMemberScoreListResDTO;
+import com.tavemakers.surf.domain.activity.dto.activityRecord.response.TeamMemberScoreListResDTO;
 import com.tavemakers.surf.domain.score.dto.response.MemberScoreRankingSliceResDTO;
 import com.tavemakers.surf.domain.score.dto.response.TeamScoreRankingSliceResDTO;
 import com.tavemakers.surf.domain.score.usecase.ScoreRankingUsecase;
